a bug fix and some cleanups
