+++ cas_client/src/local_client.rs
@@ -21,7 +21,8 @@ use tracing::{debug, error, info, warn};
 use utils::progress::ProgressUpdater;
 
 use crate::error::{CasClientError, Result};
-use crate::interface::{OutputProvider, ShardDedupProbe, UploadClient};
+use crate::interface::{ShardDedupProbe, UploadClient};
+use crate::output_provider::OutputProvider;
 use crate::{Client, ReconstructionClient, RegistrationClient, ShardClientInterface};
 
 pub struct LocalClient {
